Start using upstream liboqs branch that provides SQIsign integration
It'll probably be integrated into liboqs (the main branch) soonish.
For now, let's start using the branch already. There is a risk that
the branch disappears or is force-pushed to, in which case this build
will fail. In that case, perhaps I should push my copy of the branch
somewhere and/or update it otherwise.